*Updated toString() function to remove new lines *
In my original toString() function, I included new lines in
the returned string to better display the output.
It’s preferable in this case to retain the default output and
allow the user to construct their own custum string.
Also in this change, I  perform a minor fix to merge the
first two string segments of the output where before
they were separate strings.