Allow sorting (#353)
* *Creates sorted table of TonY records*

When viewing the TonY Portal page, the records are displayed in a random order. This fix displays TonY records sorted by most recent process completion date.

* *Adds custom toString() method to class JobMetadata*

When printing a JobMetadata object using the built in toString() method, a string representation of the object is printed rather than the metadata associated with each object (id, user, started, completed, etc). This fix prints out the associated values for each JobMetadata object when using the toString() method.

* fixed line length

* *Updated toString() function to remove new lines *

In my original toString() function, I included new lines in
the returned string to better display the output.
It’s preferable in this case to retain the default output and
allow the user to construct their own custum string.
Also in this change, I  perform a minor fix to merge the
first two string segments of the output where before
they were separate strings.