*Replaces CDNs with local file copies on TonY Portal*
When utilizing the TonY portal, users will need to be
online in order to enable the sorting/filtering mechanism
of the TonY portal in addition to bootstrap formatting
due to the usage of CDNs of various packages such as
bootstrap, jquery, and datatables. This update replaces
the original CDNs with local copies of the relevant
files in order for TonY portal to work offline in
addition to online.