* Fixes issue with testPlayBinary *
Encountered an issue in which assets are
not present in the classloader during
testPlayBinary, but are present during
runPlayBinary.  Will open an issue on
the Play Git repo for them to take a look.
In the meantime, this fix adds the project
directory to the classpath in order for
the assets to load during testPlay. In
addition, this commit fixes a minor
formatting issue in tableMetadata.scala.html.