Updates the order of TonY Portal records and adds UI for user sorting/filtering (#354)
* *Creates sorted table of TonY records*

When viewing the TonY Portal page, the records are displayed in a random order. This fix displays TonY records sorted by most recent process completion date.

* *Adds custom toString() method to class JobMetadata*

When printing a JobMetadata object using the built in toString() method, a string representation of the object is printed rather than the metadata associated with each object (id, user, started, completed, etc). This fix prints out the associated values for each JobMetadata object when using the toString() method.

* fixed line length

* *Updated toString() function to remove new lines *

In my original toString() function, I included new lines in
the returned string to better display the output.
It’s preferable in this case to retain the default output and
allow the user to construct their own custum string.
Also in this change, I  perform a minor fix to merge the
first two string segments of the output where before
they were separate strings.

* *Updated toString() function to remove new lines *

In my original toString() function, I included new lines in
the returned string to better display the output.
It’s preferable in this case to retain the default output and
allow the user to construct their own custum string.
Also in this change, I  perform a minor fix to merge the
first two string segments of the output where before
they were separate strings.

* *Updates TonY Portal Default View to Order by Multiple Fields*

On the TonY portal, jobs are currently ordered by completed
date.  However jobs that are in progress (without a completed date)
retain a random order.  This change fixes this issue by ordering
records first by completed date (desc), then by started date( desc),
and then by username (asc). Therefore, if multiple records have
identical completed date timestamps these records will be ordered
by start date. And if multiple records have identical start and complete
timestamps, then these will be ordered by username.

* *Adds User Sorting and Filtering Functionality to Tony Portal Page*

Currently on the TonY Portal, by default records are sorted first by
completed date, then started date, then user id. There is no user
initiated sorting or filtering functionality. This update adds the ability
to sort by column and filter by keyword.  In addition, Bootstrap is
added to enhance the table format.

* *Replaces CDNs with local file copies on TonY Portal*

When utilizing the TonY portal, users will need to be
online in order to enable the sorting/filtering mechanism
of the TonY portal in addition to bootstrap formatting
due to the usage of CDNs of various packages such as
bootstrap, jquery, and datatables. This update replaces
the original CDNs with local copies of the relevant
files in order for TonY portal to work offline in
addition to online.

* 072319 erwa review fixes

* * Fixes issue with testPlayBinary *

Encountered an issue in which assets are
not present in the classloader during
testPlayBinary, but are present during
runPlayBinary.  Will open an issue on
the Play Git repo for them to take a look.
In the meantime, this fix adds the project
directory to the classpath in order for
the assets to load during testPlay. In
addition, this commit fixes a minor
formatting issue in tableMetadata.scala.html.

* * Updated Java jdk *

Java jdk updated to openjdk8 from oraclejdk8.
Change was done to pass travis CI as per
UWFrankGu's comment.