Portal updates (#399)
* *Updates the .jhist File Path for Killed Applications*

When an application is killed by the user, there is no opportunity for a graceful shutdown and those files are never renamed from .jhist.inprogress to the final application status such as  -FAILED.jhist or -SUCCEEDED.jhist causing those applications to be interpreted by TonY as currently RUNNING on TonY portal even though they have been terminated.  This fix renames those files to -KILLED.jhist, correctly identifying the final status of those killed applications.

* *Updates TonY Portal page view*

This fix adds a parameter to the data table to maintain the default ordering of the TonY jobs and increases the default number of records from 10 to 25.

* *Added PR fixes*

* *Refactored code for tests*

Tests not passing due to issues with yarn environment.
Refactored code to eliminate creating separate yarn client
object in test class.

* *Fixed code in HistoryFileMoverTest.java*

* *Adds Yarn Env to HistoryFileMoverTest.java*

* **Adds Mocks to HistoryFileMoverTest**

This update adds a yarnclient mock to the HistoryFileMoverTest
in order to bypass the yarn requirements of the main
HistoryFileMover class. This update also adjusts the yarn client
initializetion of the main class to once when the mover starts
and not each time the killedAppIds function is called.