Handle untracked job type crash by failing the application.
There have been cases where if the 'ps' crashes, then the worker
processes hang forever waiting for establishment of session with the
ps.

This change kills the application if the ps or any of the untracked
processes exits in a FAILED state. This excludes cases where ps
exits with an exit code of 0 or was killed explicitly by the AM.