Upload python binary path if it exists when venv is not provided
This can be used to support other python packaging formats like
pex or shiv. While configuring a TonyJob, one can ommit the python_venv
and put the path to shiv or pex executable in python_binary_path. Tony
will upload that executable to hdfs and will generate the task command
correctly to use that path.