Binding to the reusable port without accepting any inbound connection (#465)
When TonY sets up the port with PORT_REUSE, it was using Netty whose default behavior is to bind to that port and starts a channel listening to the port(#456). This could undesirably cause other remote TensorFlow workers to connect to the port within TonY process.

The correct way is to just bind to that port but not accept any connection to it.
Since Netty doesn't not allow port binding only, and Java 8 which TonY is running with doesn't have native support for port reuse, we are using python to achieve this.